backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==13.1
sqlalchemy==2.0.36
alembic==1.14.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.17
numpy==2.2.0
reportlab==4.2.5
httpx==0.28.0
aiosqlite==0.20.0
